Refactor Languages.
Added file containing all available locales so languages can also be loaded from jar files.
Workaround til a better solution is found.
