+++ src/main/java/org/openbase/bco/bcozy/util/Languages.java
@@ -1,15 +1,11 @@
 package org.openbase.bco.bcozy.util;
 
-import org.openbase.jul.exception.printer.ExceptionPrinter;
 import org.slf4j.Logger;
 
 import javax.annotation.Nonnull;
-import java.io.*;
-import java.net.URL;
-import java.util.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
+import java.util.List;
+import java.util.Locale;
+import java.util.Objects;
 
 /**
  * Class to hold available languages.
@@ -46,87 +42,7 @@ public final class Languages {
      * Initializes Languages from available property-files.
      */
     Languages() {
-        List<Locale> bundles = getLocalesFromBundles();
-        this.languages = bundles.stream().map(Language::new).collect(Collectors.toList());
-    }
-
-    /**
-     * Creates Locales from each property-file available in "/languages".
-     */
-    private List<Locale> getLocalesFromBundles() {
-        List<String> matches = getFileMatches();
-        List<Locale> locales = new ArrayList<>(matches.size());
-        for (String match : matches) {
-            if (match.contains("_")) {
-                locales.add(new Locale(match.split("_")[0], match.split("_")[1]));
-            } else {
-                locales.add(new Locale(match));
-            }
-        }
-        return locales;
-    }
-
-    /**
-     * Gets all Language-Identifier (eg 'de', 'de_DE'), which have available property-files in "/languages".
-     */
-    @Nonnull
-    private List<String> getFileMatches() {
-        List<String> matches = new LinkedList<>();
-        Pattern pattern = Pattern.compile("languages_(\\w+[_\\w+])\\.properties");
-
-        final List<String> files = getResourceFolderFiles("languages");
-        if (files.isEmpty()) {
-            files.addAll(getResourceFolderFilesWithStream("languages"));
-        }
-
-        for (String file : getResourceFolderFiles("languages")) {
-            Matcher matcher = pattern.matcher(file);
-            if (matcher.find()) {
-                matches.add(matcher.group(1));
-            }
-        }
-        return matches;
-    }
-
-    /**
-     * Gets all File-Names from an Resource-Directory.
-     */
-    @Nonnull
-    private List<String> getResourceFolderFiles(String folder) {
-        ClassLoader loader = Languages.class.getClassLoader();
-        URL url = loader.getResource(folder);
-
-        File[] files = Optional.ofNullable(url)
-                .map(URL::getPath)
-                .map(File::new)
-                .map(File::listFiles)
-                .orElse(new File[0]);
-
-        List<String> names = new ArrayList<>();
-
-        for (File file : files) {
-            names.add(file.getName());
-        }
-        return names;
-    }
-
-    /**
-     * Gets all File-Names from an Resource-Directory, alternative Way
-     */
-    private List<String> getResourceFolderFilesWithStream(String folder) {
-        List<String> filenames = new ArrayList<>();
-
-        try (InputStream in = this.getClass().getClassLoader().getResourceAsStream(folder);
-             BufferedReader br = new BufferedReader(new InputStreamReader(in))) {
-
-            for (String resource = br.readLine(); resource != null; resource = br.readLine()) {
-                filenames.add(resource);
-            }
-        } catch (IOException ex) {
-            ExceptionPrinter.printHistory(ex, LOGGER);
-        }
-
-        return filenames;
+        this.languages = new FileLanguagesDetector().getLanguages();
     }
 
     /**
