+++ src/main/java/org/openbase/bco/bcozy/util/LanguagesDetector.java
@@ -1,0 +1,13 @@
+package org.openbase.bco.bcozy.util;
+
+import java.util.List;
+
+/**
+ * @author vdasilva
+ */
+public interface LanguagesDetector {
+
+    List<Language> getLanguages();
+
+
+}
