+++ src/main/resources/languages/available_languages
@@ -1,0 +1,2 @@
+de_DE
+en_US
